requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
tensorflow>=2.13.0
mediapipe>=0.10.5
Pillow>=10.0.0
